fixed folder structure for products and added auth for login
