+++ my-app/package.json
@@ -1,46 +1,47 @@
 {
-  "name": "my-app",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^5.11.10",
-    "@testing-library/react": "^11.2.6",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.21.1",
-    "faker": "^5.5.3",
-    "history": "^5.0.0",
-    "mirage": "0.0.7",
-    "miragejs": "^0.1.41",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-fontawesome": "^1.7.1",
-    "react-icons": "^4.2.0",
-    "react-router-dom": "^6.0.0-beta.0",
-    "react-scripts": "4.0.3",
-    "web-vitals": "^1.1.1"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
+   "name": "my-app",
+   "version": "0.1.0",
+   "private": true,
+   "dependencies": {
+      "@testing-library/jest-dom": "^5.11.10",
+      "@testing-library/react": "^11.2.6",
+      "@testing-library/user-event": "^12.8.3",
+      "axios": "^0.21.1",
+      "faker": "^5.5.3",
+      "history": "^5.0.0",
+      "mirage": "0.0.7",
+      "miragejs": "^0.1.41",
+      "react": "^17.0.2",
+      "react-dom": "^17.0.2",
+      "react-fontawesome": "^1.7.1",
+      "react-icons": "^4.2.0",
+      "react-router-dom": "^6.0.0-beta.0",
+      "react-scripts": "4.0.3",
+      "react-toastify": "^7.0.4",
+      "web-vitals": "^1.1.1"
+   },
+   "scripts": {
+      "start": "react-scripts start",
+      "build": "react-scripts build",
+      "test": "react-scripts test",
+      "eject": "react-scripts eject"
+   },
+   "eslintConfig": {
+      "extends": [
+         "react-app",
+         "react-app/jest"
+      ]
+   },
+   "browserslist": {
+      "production": [
+         ">0.2%",
+         "not dead",
+         "not op_mini all"
+      ],
+      "development": [
+         "last 1 chrome version",
+         "last 1 firefox version",
+         "last 1 safari version"
+      ]
+   }
 }
