+++ my-app/src/Pages/Homepage/Homepage.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export function Homepage() {
+  return <div>Homepage</div>;
+}
