+++ my-app/src/Pages/Login/Login.css
@@ -1,0 +1,4 @@
+.input {
+  width: 50vw;
+  height: 50vh;
+}
