+++ my-app/src/Pages/Register/Register.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export function Register() {
+  return <div>wasdwads</div>;
+}
