--- my-app/src/components/products.css
@@ -1,80 +1,0 @@
-@import "https://tetherui.netlify.app/card/card.css";
-
-.card-parent {
-   display: flex;
-   flex-wrap: wrap;
-
-   /* align-items: center; */
-   margin: 2rem;
-   margin-left: 16%;
-   justify-content: center;
-   padding-top: 2rem;
-   background-color: #f3f4f6;
-   margin-right: 0;
-   margin-top: 12vh;
-}
-
-.cardSecondary {
-   width: 20rem;
-   margin: 2rem;
-   background-color: white;
-}
-
-h2 {
-   align-self: center;
-   margin-bottom: 1rem;
-   color: var(--primary-color);
-}
-
-.aside {
-   background-color: #f3f4f6;
-   width: 15%;
-   padding: 1rem;
-   position: fixed;
-   height: 90vh;
-   top: 12vh;
-   display: flex;
-   flex-direction: column;
-}
-
-.productDiv {
-   display: flex;
-   /* width: 35vw; */
-}
-
-.productPrice {
-   overflow-y: hidden;
-}
-
-.reset-button {
-   display: inline-block;
-   width: 5vw;
-   align-self: center;
-   padding: 0.5rem 0;
-   background-color: var(--primary-color);
-   border-radius: 4px;
-   color: white;
-   border: none;
-   cursor: pointer;
-}
-
-.radio-buttons {
-   cursor: pointer;
-}
-
-.radio-label {
-   font-weight: 700;
-}
-
-.price-button {
-   cursor: pointer;
-}
-
-.product-title {
-   text-decoration: none;
-   color: var(--primary-color);
-}
-
-label {
-   color: black;
-}
