+++ my-app/src/contexts/AuthContext.js
@@ -1,20 +1,19 @@
-import React from "react";
-import { useState, createContext, useContext, useReducer } from "react";
+import { useState, createContext, useContext } from "react";
 
 export const AuthContext = createContext();
 
 export const AuthProvider = ({ children }) => {
-   const { auth, setAuth } = useState(
-      JSON.parse(localStorage.getItem("auth")) || ""
-   );
+  const { auth, setAuth } = useState(
+    JSON.parse(localStorage.getItem("auth")) || ""
+  );
 
-   return (
-      <AuthContext.Provider value={{ auth, setAuth }}>
-         {children}
-      </AuthContext.Provider>
-   );
+  return (
+    <AuthContext.Provider value={{ auth, setAuth }}>
+      {children}
+    </AuthContext.Provider>
+  );
 };
 
 export function useAuth() {
-   return useContext(AuthContext);
+  return useContext(AuthContext);
 }
